Use findAssigned datamodels endpoint
