+++ src/interfaces/batch-export-node-transport.ts
@@ -1,14 +1,14 @@
 import {
     ContentNodeTransport, DataModelTransport,
-    PackageDependencyTransport, PackageHistoryTransport,
-    VariablesAssignments
-} from "../api/package-manager-api";
+    PackageDependencyTransport,
+    PackageHistoryTransport, StudioDataModelTransport, VariablesAssignments
+} from "./package-manager.interfaces";
 
 export interface BatchExportNodeTransport extends ContentNodeTransport {
     workingDraftId: string;
     activatedDraftId: string;
     version?: PackageHistoryTransport;
     dependencies?: PackageDependencyTransport[];
-    datamodels?: DataModelTransport[];
+    datamodels?: StudioDataModelTransport[];
     variables?: VariablesAssignments[];
 }
