+++ src/services/package-manager/datamodel-service.ts
@@ -1,5 +1,6 @@
 import {BatchExportNodeTransport} from "../../interfaces/batch-export-node-transport";
-import {DataModelTransport, packageManagerApi} from "../../api/package-manager-api";
+import {DataModelTransport} from "../../interfaces/package-manager.interfaces";
+import {packageApi} from "../../api/package-api";
 
 class DatamodelService {
 
@@ -9,10 +10,8 @@ class DatamodelService {
         const promises = [];
 
         nodesListToExport.forEach(node => {
-            const datamodelIds: string[] = node.variables.filter(variable => variable.type === "DATA_MODEL").map(variable => variable.value as unknown as string);
-
             promises.push(new Promise(async resolve => {
-                node.datamodels = await this.getDataDatamodelsByIds(node.key, datamodelIds);
+                node.datamodels = await packageApi.findAssignedDatamodels(node.key);
                 resolve(node);
             }));
         })
@@ -20,25 +19,6 @@ class DatamodelService {
         return Promise.all(promises);
     }
 
-    private async getDataDatamodelsByIds(packageKey: string, datamodelIds: string[]): Promise<DataModelTransport[]> {
-        const promises = [];
-
-        datamodelIds.forEach(datamodelId => {
-            promises.push(new Promise(async resolve => {
-                if (datamodelId) {
-                    let datamodel;
-                    try {
-                        datamodel = await packageManagerApi.getDataModel(datamodelId, packageKey);
-                    } catch (e) {
-                        resolve();
-                    }
-                    resolve(datamodel);
-                }
-            }))
-        })
-
-        return Promise.all(promises);
-    }
 }
 
 export const dataModelService = DatamodelService.INSTANCE;
